+++ PHPUnit/Util/Configuration.php
@@ -139,6 +139,7 @@
  *     <log type="junit" target="/tmp/logfile.xml" logIncompleteSkipped="false"/>
  *     <log type="testdox-html" target="/tmp/testdox.html"/>
  *     <log type="testdox-text" target="/tmp/testdox.txt"/>
+ *     <log type="coverage-crap4j" target="/tmp/crap.xml"/>
  *   </logging>
  *
  *   <php>
@@ -541,10 +542,23 @@ class PHPUnit_Util_Configuration
         }
 
         foreach (array('var', 'env', 'post', 'get', 'cookie', 'server', 'files', 'request') as $array) {
-            if ($array == 'var') {
-                $target = &$GLOBALS;
-            } else {
-                $target = &$GLOBALS['_' . strtoupper($array)];
+            // See https://github.com/sebastianbergmann/phpunit/issues/277
+            switch ($array) {
+                case 'var':
+                    $target = &$GLOBALS;
+                    break;
+
+                case 'env':
+                    $target = &$_ENV;
+                    break;
+
+                case 'server':
+                    $target = &$_SERVER;
+                    break;
+
+                default:
+                    $target = &$GLOBALS['_' . strtoupper($array)];
+                    break;
             }
 
             foreach ($configuration[$array] as $name => $value) {