+++ PHPUnit/Util/ErrorHandler.php
@@ -45,10 +45,10 @@
 
 // Workaround for http://bugs.php.net/bug.php?id=47987,
 // see https://github.com/sebastianbergmann/phpunit/issues#issue/125 for details
-require_once 'PHPUnit/Framework/Error.php';
-require_once 'PHPUnit/Framework/Error/Notice.php';
-require_once 'PHPUnit/Framework/Error/Warning.php';
-require_once 'PHPUnit/Framework/Error/Deprecated.php';
+require_once __DIR__ . '/../Framework/Error.php';
+require_once __DIR__ . '/../Framework/Error/Notice.php';
+require_once __DIR__ . '/../Framework/Error/Warning.php';
+require_once __DIR__ . '/../Framework/Error/Deprecated.php';
 
 /**
  * Error handler that converts PHP errors and warnings to exceptions.