+++ Tests/Framework/ConstraintTest.php
@@ -3023,6 +3023,24 @@ EOF
 
     /**
      * @covers PHPUnit_Framework_Constraint_TraversableContains
+     */
+    public function testConstraintArrayContainsCheckForObjectIdentity()
+    {
+        // Check for primitive type.
+        $constraint = new PHPUnit_Framework_Constraint_TraversableContains('foo', TRUE, TRUE);
+
+        $this->assertFalse($constraint->evaluate(array(0), '', TRUE));
+        $this->assertFalse($constraint->evaluate(array(TRUE), '', TRUE));
+
+        // Default case.
+        $constraint = new PHPUnit_Framework_Constraint_TraversableContains('foo');
+
+        $this->assertTrue($constraint->evaluate(array(0), '', TRUE));
+        $this->assertTrue($constraint->evaluate(array(TRUE), '', TRUE));
+    }
+
+    /**
+     * @covers PHPUnit_Framework_Constraint_TraversableContains
      * @covers PHPUnit_Framework_Constraint::count
      * @covers PHPUnit_Framework_TestFailure::exceptionToString
      */
@@ -3082,24 +3100,6 @@ EOF
         }
 
         $this->fail();
-    }
-
-    /**
-     * @covers PHPUnit_Framework_Constraint_TraversableContains
-     */
-    public function testConstraintArrayContainsCheckForObjectIdentity()
-    {
-        // Check for primitive type.
-        $constraint = new PHPUnit_Framework_Constraint_TraversableContains('foo', TRUE, TRUE);
-
-        $this->assertFalse($constraint->evaluate(array(0), '', TRUE));
-        $this->assertFalse($constraint->evaluate(array(TRUE), '', TRUE));
-
-        // Default case.
-        $constraint = new PHPUnit_Framework_Constraint_TraversableContains('foo');
-
-        $this->assertTrue($constraint->evaluate(array(0), '', TRUE));
-        $this->assertTrue($constraint->evaluate(array(TRUE), '', TRUE));
     }
 
     /**
@@ -3522,6 +3522,99 @@ EOF
 
         try {
             $constraint->evaluate(array(1,2));
+        }
+
+        catch (PHPUnit_Framework_ExpectationFailedException $e) {
+            $this->assertEquals(
+              <<<EOF
+Failed asserting that actual size 2 does not match expected size 2.
+
+EOF
+              ,
+              PHPUnit_Framework_TestFailure::exceptionToString($e)
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+
+    /**
+     * @covers PHPUnit_Framework_Constraint_SameSize
+     */
+    public function testConstraintSameSizeWithAnArray()
+    {
+        $constraint = new PHPUnit_Framework_Constraint_SameSize(array(1,2,3,4,5));
+
+        $this->assertTrue($constraint->evaluate(array(6,7,8,9,10), '', TRUE));
+        $this->assertFalse($constraint->evaluate(array(1,2,3,4), '', TRUE));
+    }
+
+    /**
+     * @covers PHPUnit_Framework_Constraint_SameSize
+     */
+    public function testConstraintSameSizeWithAnIteratorWhichDoesNotImplementCountable()
+    {
+        $constraint = new PHPUnit_Framework_Constraint_SameSize(new TestIterator(array(1,2,3,4,5)));
+
+        $this->assertTrue($constraint->evaluate(new TestIterator(array(6,7,8,9,10)), '', TRUE));
+        $this->assertFalse($constraint->evaluate(new TestIterator(array(1,2,3,4)), '', TRUE));
+    }
+
+    /**
+     * @covers PHPUnit_Framework_Constraint_SameSize
+     */
+    public function testConstraintSameSizeWithAnObjectImplementingCountable()
+    {
+        $constraint = new PHPUnit_Framework_Constraint_SameSize(new ArrayObject(array(1,2,3,4,5)));
+
+        $this->assertTrue($constraint->evaluate(new ArrayObject(array(6,7,8,9,10)), '', TRUE));
+        $this->assertFalse($constraint->evaluate(new ArrayObject(array(1,2,3,4)), '', TRUE));
+    }
+
+    /**
+     * @covers PHPUnit_Framework_Constraint_SameSize
+     * @covers PHPUnit_Framework_TestFailure::exceptionToString
+     */
+    public function testConstraintSameSizeFailing()
+    {
+        $constraint = new PHPUnit_Framework_Constraint_SameSize(array(1,2,3,4,5));
+
+        try {
+            $constraint->evaluate(array(1,2));
+        }
+
+        catch (PHPUnit_Framework_ExpectationFailedException $e) {
+            $this->assertEquals(
+              <<<EOF
+Failed asserting that actual size 2 matches expected size 5.
+
+EOF
+              ,
+              PHPUnit_Framework_TestFailure::exceptionToString($e)
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+
+    /**
+     * @covers PHPUnit_Framework_Constraint_SameSize
+     * @covers PHPUnit_Framework_Constraint_Not
+     * @covers PHPUnit_Framework_Assert::logicalNot
+     * @covers PHPUnit_Framework_TestFailure::exceptionToString
+     */
+    public function testConstraintNotSameSizeFailing()
+    {
+        $constraint = PHPUnit_Framework_Assert::logicalNot(
+          new PHPUnit_Framework_Constraint_SameSize(array(1,2))
+        );
+
+        try {
+            $constraint->evaluate(array(3,4));
         }
 
         catch (PHPUnit_Framework_ExpectationFailedException $e) {